[refactor] 패키지 위치 이동
일관성을 위해 패키지 위치를 변경합니다
